move ThrowableThrownFailureHandler to domain-api
moreover:
- add factory methods to SubjectChanger.FailureHandler
- rename DefaultFailureHandlerImpl to DefaultFailureHandler
- make FailureHandlerAdapter internal
