+++ domain/api/atrium-domain-api-common/src/main/kotlin/ch/tutteli/atrium/domain/creating/changers/impl/subjectchanger/DefaultFailureHandler.kt
@@ -1,0 +1,31 @@
+package ch.tutteli.atrium.domain.creating.changers.impl.subjectchanger
+
+import ch.tutteli.atrium.assertions.Assertion
+import ch.tutteli.atrium.assertions.builders.assertionBuilder
+import ch.tutteli.atrium.assertions.builders.invisibleGroup
+import ch.tutteli.atrium.core.None
+import ch.tutteli.atrium.core.Option
+import ch.tutteli.atrium.creating.Expect
+import ch.tutteli.atrium.domain.creating.changers.SubjectChanger
+import ch.tutteli.atrium.domain.creating.collectors.collectAssertions
+
+class DefaultFailureHandler<T, R> : SubjectChanger.FailureHandler<T, R> {
+
+    override fun createAssertion(
+        originalAssertionContainer: Expect<T>,
+        descriptiveAssertion: Assertion,
+        maybeAssertionCreator: Option<Expect<R>.() -> Unit>
+    ): Assertion = maybeAssertionCreator.fold({
+        descriptiveAssertion
+    }) { assertionCreator ->
+        assertionBuilder.invisibleGroup
+            .withAssertions(
+                descriptiveAssertion,
+                assertionBuilder.explanatoryGroup
+                    .withDefaultType
+                    .collectAssertions(None, assertionCreator)
+                    .build()
+            )
+            .build()
+    }
+}
