+++ domain/api/atrium-domain-api-common/src/main/kotlin/ch/tutteli/atrium/domain/creating/changers/impl/subjectchanger/ThrowableThrownFailureHandler.kt
@@ -7,25 +7,36 @@ import ch.tutteli.atrium.core.None
 import ch.tutteli.atrium.core.Option
 import ch.tutteli.atrium.creating.Expect
 import ch.tutteli.atrium.domain.creating.changers.SubjectChanger
+import ch.tutteli.atrium.domain.creating.changers.utils.propertiesOfThrowable
 import ch.tutteli.atrium.domain.creating.collectors.collectAssertions
 
-class DefaultFailureHandlerImpl<T, R> : SubjectChanger.FailureHandler<T, R> {
+class ThrowableThrownFailureHandler<T : Throwable?, R>(
+    private val maxStackTrace: Int
+) : SubjectChanger.FailureHandler<T, R> {
 
     override fun createAssertion(
         originalAssertionContainer: Expect<T>,
         descriptiveAssertion: Assertion,
         maybeAssertionCreator: Option<Expect<R>.() -> Unit>
-    ): Assertion = maybeAssertionCreator.fold({
-        descriptiveAssertion
-    }) { assertionCreator ->
-        assertionBuilder.invisibleGroup
-            .withAssertions(
-                descriptiveAssertion,
+    ): Assertion {
+        val assertions = mutableListOf(descriptiveAssertion)
+        maybeAssertionCreator.fold({ /* nothing to do */ }) { assertionCreator ->
+            assertions.add(
                 assertionBuilder.explanatoryGroup
                     .withDefaultType
                     .collectAssertions(None, assertionCreator)
                     .build()
             )
+        }
+        originalAssertionContainer.maybeSubject.fold(
+            { /* nothing to do */ },
+            {
+                if (it != null) assertions.add(propertiesOfThrowable(it, maxStackTrace))
+            }
+        )
+        return assertionBuilder.invisibleGroup
+            .withAssertions(assertions.toList())
             .build()
     }
 }
+
