+++ domain/api/atrium-domain-api-jvm/src/main/kotlin/ch/tutteli/atrium/domain/creating/changers/utils/createAdditionalHints.kt
@@ -1,6 +1,6 @@
 @file:Suppress("JAVA_MODULE_DOES_NOT_READ_UNNAMED_MODULE" /* TODO remove once https://youtrack.jetbrains.com/issue/KT-35343 is fixed */)
 
-package ch.tutteli.atrium.domain.robstoll.lib.creating.throwable.thrown.creators
+package ch.tutteli.atrium.domain.creating.changers.utils
 
 import ch.tutteli.atrium.assertions.Assertion
 import ch.tutteli.atrium.translations.DescriptionThrowableAssertion
@@ -9,7 +9,7 @@ actual fun createAdditionalHints(
     throwable: Throwable,
     maxStackTrace: Int
 ): List<Assertion> = throwable.suppressed.map { suppressed ->
-    ThrowableThrownFailureHandler.createChildHint(
+    createChildHint(
         throwable,
         suppressed,
         DescriptionThrowableAssertion.OCCURRED_EXCEPTION_SUPPRESSED,
