+++ domain/robstoll-lib/atrium-domain-robstoll-lib-jvm/src/main/kotlin/ch/tutteli/atrium/domain/robstoll/lib/creating/pathAssertions.kt
@@ -9,11 +9,10 @@ import ch.tutteli.atrium.core.None
 import ch.tutteli.atrium.core.Some
 import ch.tutteli.atrium.core.polyfills.fullName
 import ch.tutteli.atrium.creating.Expect
-import ch.tutteli.atrium.domain.builders.AssertImpl
 import ch.tutteli.atrium.domain.creating._domain
 import ch.tutteli.atrium.domain.creating.changers.ExtractedFeaturePostStep
+import ch.tutteli.atrium.domain.creating.changers.utils.propertiesOfThrowable
 import ch.tutteli.atrium.domain.robstoll.lib.creating.filesystem.*
-import ch.tutteli.atrium.domain.robstoll.lib.creating.throwable.thrown.creators.ThrowableThrownFailureHandler
 import ch.tutteli.atrium.reporting.RawString
 import ch.tutteli.atrium.reporting.translating.Translatable
 import ch.tutteli.atrium.translations.DescriptionBasic.*
@@ -357,10 +356,10 @@ private fun hintForNotDirectory(actualType: Translatable) =
         .build()
 
 private fun hintForOtherIoException(exception: IOException) =
-    ThrowableThrownFailureHandler.propertiesOfThrowable(
+    propertiesOfThrowable(
         exception,
         maxStackTrace = IO_EXCEPTION_STACK_TRACE_LENGTH,
-        explanation = AssertImpl.builder.explanatory
+        explanation = assertionBuilder.explanatory
             .withExplanation(
                 FAILURE_DUE_TO_ACCESS_EXCEPTION,
                 exception::class.simpleName ?: exception::class.fullName
